feat: implement forget API, switch to int64 types, and add a context manager for the stdout logging wrapping for libonlinesvr

+++ pyonlinesvr/lib/compat.py
@@ -60,7 +60,7 @@ def double_matrix_to_np(m: DoubleMatrix) -> np.ndarray:
 
 
 def np_to_int_vector(arr: np.ndarray) -> IntVector:
-    arr = check_array(arr, dtype=np.int8, ensure_2d=False, order="C", copy=False,
+    arr = check_array(arr, dtype=np.int64, ensure_2d=False, order="C", copy=False,
                       accept_sparse=False, accept_large_sparse=False)
     if len(arr.shape) > 1:
         raise ValueError("Can only convert a 1-dimensional array into a IntVector! "
@@ -68,20 +68,20 @@ def np_to_int_vector(arr: np.ndarray) -> IntVector:
 
     vec = IntVector()
     for v in arr:
-        vec.Add(v)
+        vec.Add(int(v))
     return vec
 
 
 def int_vector_to_np(vec: IntVector) -> np.ndarray:
     length = vec.GetLength()
-    arr = np.empty(length, dtype=np.int8, order="C")
+    arr = np.empty(length, dtype=np.int64, order="C")
     for i in range(length):
         arr[i] = vec.GetValue(i)
     return arr
 
 
 def np_to_int_matrix(arr: np.ndarray) -> IntMatrix:
-    arr = check_array(arr, dtype=np.int8, ensure_2d=True, order="C", copy=False,
+    arr = check_array(arr, dtype=np.int64, ensure_2d=True, order="C", copy=False,
                       accept_sparse=False, accept_large_sparse=False)
     m = IntMatrix()
     for row in arr:
@@ -94,7 +94,7 @@ def int_matrix_to_np(m: IntMatrix) -> np.ndarray:
     rows = m.GetLengthRows()
     cols = m.GetLengthCols()
 
-    arr = np.empty((rows, cols), dtype=np.int8, order="C")
+    arr = np.empty((rows, cols), dtype=np.int64, order="C")
     for i in range(rows):
         for j in range(cols):
             arr[i, j] = m.GetValue(i, j)
